Spring Web vulnerable to Open Redirect or Server Side Request Forgery (#39)
Package
Affected versions
Patched version
org.springframework:spring-web
(Maven)
>= 5.3.0, < 5.3.32
5.3.32
Applications that use UriComponentsBuilder to parse an externally provided URL (e.g. through a query parameter) AND perform validation checks on the host of the parsed URL may be vulnerable to a open redirect attack or to a SSRF attack if the URL is used after passing validation checks.